Get rid of compiler errors for util_i_library.nss
These errors are an artifact of the implementation, not actual errors.
util_i_library.nss should not compile on its own since (even though it
implements main()) it is included in library files.

Silencing these errors involves using the nwnsc #pragma compiler
extension, so it must be compiled with the -e flag. This cannot be done
in the toolset, causing library compilation to fail. To compile in the
toolset, comment out the offending lines in util_i_library.nss.