refactor!: use glob matching to load libraries
This removes the TestStringAgainstPattern() version of
LoadLibrariesByPattern() and replaces it with glob syntax. The function
also takes a CSV list of patterns, and loads the libraries in the order
they match the patterns, allowing finer control over the load order.

However, this is a BREAKING CHANGE. Sorry if you built your module on
that during the whole day the previous version was up.