feat: add times utility (#8)
This is a utility for handling dates, times, durations, and the
formatting thereof.

The formatting functions in utl_i_strftime are largely a
re-implementation of C's strftime() in NWScript. The format codes should
be familiar to C and python devs as well as anyone who uses the date
utility in Linux. For simple usage, the TimeToString() function is
provided by util_i_times.

Thanks to @tinygiant98 for his thorough testing!