feat: add utility `util_i_argstack`
Add `util_i_argstack` to allow more robust passage of variables between functions when functions are not inclusive to the current script. This primarily targets the library system and is similar to, but a much expanded version of, script parameters.

Each type includes a function to: `Push`, `Pop`, `Peek` and `Count` the type's argument stack.