style: drive-by formatting and general administrative update (#40)
Noticed a couple of errors in a file I was working with, so went through most of the files for some basic updates to administrative data (comments, documents, etc.). Added deprecation warning debug messages to old functions that should be removed. No functionality has been added or changed with this commit.